[TECH-556] Move yaml utility methods to their own package
branch: feature/TECH-556-config-upgrades
